+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
-
+### Changed
+- [Patch] Update input story to include disabled state
 
 ## 31.12.4 - 2018-09-10
 ### Changed